% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromJSONL.R
\name{fromJSONL}
\alias{fromJSONL}
\title{fromJSONL}
\usage{
fromJSONL(filepath, maxLines = -1)
}
\arguments{
\item{filepath}{A character string indicating the filepath from the working
directory to the desired file.}

\item{maxLines}{An integer representing the max number of lines to read. Negative
values indicate that one should read up to the end of input on the connection.}
}
\value{
a large list, each element containg the contents of a JSON file after
   being converted.
}
\description{
Decode a JSON List into an R List Object.
}
