% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure.R
\name{getStructure}
\alias{getStructure}
\title{getStructure}
\usage{
getStructure(protein)
}
\arguments{
\item{protein}{input for a a protein object}
}
\value{
Returns a dataframe containing the atomic record. There are 15 variables
 in this data frame.
}
\description{
Retrieve the Atomic Record from a Protein Object
}
\details{
This is an accessor function for retrieving the Atomic Record from a
 Protein object.
}
