% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTitleSection.R
\name{getTitleSection}
\alias{getTitleSection}
\title{getTitleSection}
\usage{
getTitleSection(protein)
}
\arguments{
\item{protein}{input for a a protein object}
}
\value{
Returns a list containing elements from the title section.
}
\description{
Retrieve the title section from a Protein Object
}
\details{
This is an accessor function for retrieving the title section from a
 Protein object.
}
