% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_write.R
\name{image_write}
\alias{image_write}
\title{image_write}
\usage{
image_write(
  image,
  path = NULL,
  format = NULL,
  quality = NULL,
  depth = NULL,
  density = NULL,
  comment = NULL,
  flatten = FALSE,
  defines = NULL
)
}
\arguments{
\item{image}{magick image object.}

\item{path}{a file, url, or raster object or bitmap array}

\item{format}{output format such as "png", "jpeg", "gif", "rgb" or "rgba".}

\item{quality}{number between 0 and 100 for jpeg quality. Defaults to 75.}

\item{depth}{color depth (either 8 or 16)}

\item{density}{resolution to render pdf or svg}

\item{comment}{text string added to the image metadata for supported formats.}

\item{flatten}{should image be flattened before writing? This also replaces transparency with background color.}

\item{defines}{a named character vector with extra options to control reading. These are the  -define key{=value} settings in the command line tool. Use an empty string for value-less defines, and NA to unset a define.}
}
\description{
Wrapper function for writing images to the disk. This function comes from the
magick package under the same name.
}
