% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProtein3D.R
\name{plot3D}
\alias{plot3D}
\title{plot3D}
\usage{
plot3D(protein, animated = FALSE, groups = NULL)
}
\arguments{
\item{protein}{Protein object to be plotted}

\item{animated}{logical indicating whether the object is to be animated in the
viewer.}

\item{groups}{indicate which groups to color. This argument is still in development
and does not currently work.}
}
\value{
An object to be plotted. If not assigned to a variable, it will plot
 directly in the viewer.
}
\description{
plot the protein structure in 3D
}
\details{
This function uses lattice and magick to create the 3D plot and animate
 it.

Currently this function is incomplete and will change dramatically as
 new features and documentation are added.
}
