% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModels.R
\name{plotModels}
\alias{plotModels}
\title{plotModels}
\usage{
plotModels(protein, separate = FALSE)
}
\arguments{
\item{protein}{Protein object to be plotted}

\item{separate}{indicate wether to plot each plane separately or as one visual.}
}
\value{
An object to be plotted. If not assigned to a variable, it will plot
 directly in the viewer.
}
\description{
plot models of the protein structure using ggplot.
}
\details{
Create a plot of each plane and model the shape of the protein.

This function uses ggplot and grid to create 3 plots, one for each plane,
 of the protein model, and then create a smoothing model.


Currently this function is incomplete and will change dramatically as
 new features and documentation are added.
}
