% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinClassDefinition.R
\docType{class}
\name{Protein-class}
\alias{Protein-class}
\title{Protein Class Definitions
Is used to Define Protein Objects of S3 and S4 Types}
\description{
Protein Class Definitions
Is used to Define Protein Objects of S3 and S4 Types
}
