% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3DInteractiveShinyApp.R
\name{Protein3D}
\alias{Protein3D}
\title{plot3DInteractive}
\usage{
Protein3D(protein)
}
\arguments{
\item{protein}{Protein object to use in plotting}
}
\description{
Opens a viewer for exploratory and interactive analysis of a protein structure.
}
