% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_dataframe.R
\name{report_as_dataframe}
\alias{report_as_dataframe}
\title{report_as_dataframe}
\usage{
report_as_dataframe(report, records = c(1:length(report)))
}
\arguments{
\item{report}{a list derived a vaccine report from NCBI Datasets.}

\item{records}{a vector of indices to pull from the report.}
}
\value{
A large dataframe with 23 variables containig metadata from NCBI Virus
 report.
}
\description{
Function to transform a list of NCBI Virus Report metadata into a table.
}
