% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_viz.R
\name{write_viz}
\alias{write_viz}
\title{write_viz}
\usage{
write_viz(image, path = "my_image", format = "png")
}
\arguments{
\item{image}{magick image object or trellis object.}

\item{path}{a file path starting from the working directory}

\item{format}{file type to save the image as. Can be "png", "jpeg", "gif", "rgb",
or "rgba".}
}
\description{
Wrapper function for writing images to the disk. This function comes from the
magick package under the same name.
}
