% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure.R
\name{getStructure}
\alias{getStructure}
\title{getStructure}
\usage{
getStructure(protein)
}
\arguments{
\item{protein}{input for a a protein object}
}
\description{
Retrieves Protein Structure from Protein Object
}
