% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinClassDefinition.R
\docType{class}
\name{Protein-class}
\alias{Protein-class}
\title{Protein Class Definitions}
\description{
Protein Class used to Define Protein Objects of S3 and S4 Types.
Currently still in development, Integrity checks still need to be added.
}
