% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.pdb.R
\name{read.pdb}
\alias{read.pdb}
\title{read.pdb}
\usage{
read.pdb(fileName, createAsS4 = TRUE)
}
\arguments{
\item{fileName}{character string for location and name of file to be read.}

\item{createAsS4}{Logical indicating whether to create the new protein object
as S4 or not. Defaults to TRUE if not specified. This argument is optional.}
}
\value{
A new protein object as either an S3 or S4 object.

In general terms, the new object will be a list of two, a data frame
 containing the atomic record, and a list of header elements.
}
\description{
Read in a Protein Data Bank file
}
\details{
Reads a Protein Data Bank file (PDB) from the given location. The
 function then parses the file and creates a new object of the Protein class.
 This object can be either defined as an S3 or S4 object if different
 capabilities are required.
}
